Implemented Kafka Configurations
- Added Support For Basic Kafka Configuration
- Updated Consumer, Group, Producer Service To Use This Configurations

+++ app/interface/partition.go
@@ -4,5 +4,5 @@ import "github.com/Raman5837/kafka.go/app/types"
 
 // PartitionAssigner Interface
 type PartitionAssignerInterface interface {
-	Next(topicID uint64) (*types.GetPartition, error)
+	Next(topicID uint) (*types.GetPartition, error)
 }
